Doorstroming aanpassingen veel bugs opgelost

+++ Project/DataModels/ProductModel.cs
@@ -1,15 +1,15 @@
 public class ProductModel
 {
-    public Int64 ProductID {get; set;}
+    public int ProductID {get; set;}
     public string Name {get; set;}
-    public Decimal Price {get; set;}
+    public decimal Price {get; set;}
     public string Brand {get; set;}
     public string Ingredients {get; set;}
     public Int64 CategoryID {get; set;}
     public string Category {get; set;}
     public Int64 Stock {get; set;}
 
-    public ProductModel(Int64 productid, string name, Decimal price, string brand, string ingredients, Int64 categoryid, Int64 stock)
+    public ProductModel(int productid, string name, Decimal price, string brand, string ingredients, int categoryid, Int64 stock)
     {
         ProductID = productid;
         Name = name;
